backend/pyproject.toml
[project]
name = "auth-backend"
version = "0.1.0"
description = "Multi-subject authentication & session management system"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "alembic>=1.14",
    "argon2-cffi>=23.1",
    "redis[hiredis]>=5.0",
    "pydantic-settings>=2.7",
    "python-dotenv>=1.0",
    "user-agents>=2.2",
    "webauthn>=2.7.1",
    "pyotp>=2.9.0",
    "qrcode>=8.2",
]
